data: represent scalars as `f32` (#4395)
Summary:
In-memory and wire representations of scalars are now 32-bit rather than
64-bit floats. We no longer support reading `DT_DOUBLE` tensors, which
is fine because TensorBoard has never written such tensors to event
files. This is mostly a no-op on the Python side, except that Python
`float`-to-`float` conversions that go through the data provider layer
are now lossy.

Per a comment of @nfelt on #4352.

Test Plan:
All the Rust changes follow from the types. Tests suffice, as usual.

wchargin-branch: data-scalars-f32

+++ tensorboard/data/server/commit.rs
@@ -103,7 +103,7 @@ pub struct DataLoss;
 
 /// The value of a scalar time series at a single point.
 #[derive(Debug, Copy, Clone, PartialEq)]
-pub struct ScalarValue(pub f64);
+pub struct ScalarValue(pub f32);
 
 #[cfg(test)]
 mod tests {
